feat: ページコンポーネントのリファクタリングと新機能の追加
- Page1コンポーネントをMain、Side、SpeakerCardsに分割
- Page2コンポーネントにAvatarPreLoaderとPage3ImagePreLoaderを追加
- Page3コンポーネントを新規作成し、画像プリロード機能を実装
- Page4コンポーネントを新規作成し、動画プレイリスト機能を追加
- PageHeaderコンポーネントを新規作成し、ヘッダー表示を改善
- next.config.jsにPWA設定を追加し、画像の品質を設定
- 不要なコードの削除と整理
