+++ src/pages/break-dk/talks/[talkId].tsx
@@ -1,6 +1,9 @@
 import AudioPlayer from '@/components/media/AudioPlayer'
 import Page1 from '@/components/pages/Page1'
-import Page2, { AvatarPreLoader } from '@/components/pages/Page2'
+import Page2, {
+  AvatarPreLoader,
+  Page3ImagePreLoader,
+} from '@/components/pages/Page2'
 import Page3 from '@/components/pages/Page3'
 import Page4 from '@/components/pages/Page4'
 import Loading from '@/components/common/Loading'
@@ -99,6 +102,7 @@ function Pages() {
       />
       <AudioPlayer src={audioSrc} shouldPlay={shouldPlayAudio} />
       <AvatarPreLoader view={view} isDk={true}></AvatarPreLoader>
+      <Page3ImagePreLoader isDk={true} />
       <div className="w-[1920px] h-[1080px] relative text-black overflow-hidden">
         <Image
           src={backgroundSrc}
